Rename lib/package.js to lib/packageConfig.js.
This is to remove confusion with the reserved package variable name.
