+++ app/lib/packageConfig.js
